requirements.txt
Flask==2.1.2
flask-cors==3.1.1
torch==1.12.1
ultralytics==8.0.22
Pillow==9.1.0
opencv-python==4.6.0.66
numpy==1.23.1
requests==2.28.1